+++ src/OctoshiftCLI.Tests/Octoshift/HttpDownloadServiceFactoryTests.cs
@@ -2,6 +2,7 @@
 using FluentAssertions;
 using Moq;
 using OctoshiftCLI.Contracts;
+using OctoshiftCLI.Factories;
 using OctoshiftCLI.Services;
 using Xunit;
